functional UI

+++ src/app/grocery-list/grocery-item/grocery-item.component.ts
@@ -1,7 +1,9 @@
-// import { Component, OnInit } from '@angular/core';
-// import { ActivatedRoute } from '@angular/router';
+// import { Component, Input, OnInit } from '@angular/core';
+// import { ActivatedRoute, ParamMap } from '@angular/router';
+// import { switchMap } from 'rxjs/operators';
+// import { SharedService } from '../../shared.service';
 // import { GroceryItem, GroceryList } from '../models/grocery.models';
-// import { SharedService } from 'src/app/shared.service';
+// import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 
 // @Component({
 //   selector: 'app-grocery-item',
@@ -9,83 +11,108 @@
 //   styleUrls: ['./grocery-item.component.css'],
 // })
 // export class GroceryItemComponent implements OnInit {
-//   listId!: number;
+//   @Input() list!: GroceryList;
+//   listId = this.list?.id;
+
+//   selectedItem!: GroceryItem;
+
 //   groceryItems: GroceryItem[] = [];
-//   newItemName!: string;
-//   newItemQuantity!: number;
-//   updatingItemId!: number; // Used to track the item being updated
+//   groceryItemForm: FormGroup;
+//   // newItemName: string = '';
+//   // newItemQuantity!: number;
+//   purchasedForm: FormGroup | undefined;
 
 //   constructor(
 //     private route: ActivatedRoute,
-//     private sharedService: SharedService
-//   ) {}
-
-//   ngOnInit(): void {
-//     this.route.paramMap.subscribe((params) => {
-//       this.listId = +params.get('listId')!;
-//       this.loadGroceryItems();
+//     private groceryService: SharedService,
+//     private formBuilder: FormBuilder
+//   ) {
+//     this.groceryItemForm = this.formBuilder.group({
+//       name: ['', Validators.required],
+//       quantity: [1, Validators.required],
 //     });
 //   }
 
+//   ngOnInit(): void {
+//     this.loadGroceryItems();
+//   }
+
 //   loadGroceryItems(): void {
-//     this.sharedService
-//       .getGroceryItemsForList(this.listId)
+//     this.route.paramMap
+//       .pipe(
+//         switchMap((params: ParamMap) => {
+//           this.listId = +params.get('listId')!;
+//           return this.groceryService.getGroceryItemsForList(this.listId);
+//         })
+//       )
 //       .subscribe((items) => {
 //         this.groceryItems = items;
 //       });
 //   }
 
 //   createGroceryItem(): void {
-//     const newItem: GroceryItem = {
-//       id: 0,
-//       name: this.newItemName,
-//       quantity: this.newItemQuantity,
-//       description: null,
-//       purchased: false,
-//       grocery_list: {
-//         id: this.listId,
-//         name: '',
-//         description: '',
-//       } as GroceryList,
-//     };
-//     this.sharedService
-//       .createGroceryItemForList(this.listId, newItem)
-//       .subscribe(() => {
-//         this.loadGroceryItems();
-//         this.newItemName = ''; // Clear the input fields
-//         this.newItemQuantity = 1;
-//       });
-//   }
-
-//   updateGroceryItem(item: GroceryItem): void {
-//     // Ensure that an item is selected for updating
-//     if (!item || !item.id) {
+//     if (this.groceryItemForm.invalid) {
 //       return;
 //     }
-//     this.sharedService
-//       .updateGroceryItemForList(this.listId, item.id, item)
-//       .subscribe(() => {
-//         this.loadGroceryItems();
-//         //this.updatingItemId = null;
+
+//     const newItem: Partial<GroceryItem> = {
+//       name: this.groceryItemForm.value.name,
+//       quantity: this.groceryItemForm.value.quantity,
+//       grocery_list: this.listId,
+//     };
+
+//     this.groceryService
+//       .createGroceryItemForList(this.listId, newItem as GroceryItem)
+//       .subscribe((item) => {
+//         this.groceryItems.push(item);
+//         this.groceryItemForm.reset(); // Reset the form
+//         this.groceryItemForm.patchValue({ quantity: 1 }); // Set quantity to 1
 //       });
 //   }
 
-//   deleteGroceryItem(itemId: number): void {
-//     this.sharedService
-//       .deleteGroceryItemForList(this.listId, itemId)
+//   openModal(item: GroceryItem) {
+//     this.selectedItem = item;
+//     this.groceryItemForm.setValue({
+//       name: item.name,
+//       quantity: item.quantity,
+//       purchased: item.purchased,
+//     });
+//   }
+
+//   updateGroceryItem(): void {
+//     if (!this.selectedItem) {
+//       return;
+//     }
+
+//     // Update the selected grocery item with the edited values
+//     this.selectedItem.name = this.groceryItemForm.value.name;
+//     this.selectedItem.quantity = this.groceryItemForm.value.quantity;
+
+//     // Call the API to update the grocery item
+//     this.groceryService
+//       .updateGroceryItemForList(
+//         this.listId,
+//         this.selectedItem.id,
+//         this.selectedItem
+//       )
 //       .subscribe(() => {
+//         // After successful update, close the modal and reload the grocery items
 //         this.loadGroceryItems();
 //       });
 //   }
 
-//   // Start updating a specific item
-//   startUpdatingItem(itemId: number): void {
-//     this.updatingItemId = itemId;
+//   deleteGroceryItem(item: GroceryItem): void {
+//     this.groceryService
+//       .deleteGroceryItemForList(this.listId, item.id)
+//       .subscribe(() => {
+//         this.groceryItems = this.groceryItems.filter((i) => i.id !== item.id);
+//       });
 //   }
 
-//   // Cancel updating the item
-//   cancelUpdatingItem(): void {
-//     //this.updatingItemId = null;
+//   markAsPurchased(item: GroceryItem, event: any): void {
+//     const purchasedValue = event.target.checked;
+//     item.purchased = purchasedValue;
+//     this.updateGroceryItem();
 //   }
 // }
 
@@ -93,7 +120,8 @@ import { Component, Input, OnInit } from '@angular/core';
 import { ActivatedRoute, ParamMap } from '@angular/router';
 import { switchMap } from 'rxjs/operators';
 import { SharedService } from '../../shared.service';
-import { GroceryItem } from '../models/grocery.models';
+import { GroceryItem, GroceryList } from '../../models/grocery.models';
+import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 
 @Component({
   selector: 'app-grocery-item',
@@ -101,21 +129,41 @@ import { GroceryItem } from '../models/grocery.models';
   styleUrls: ['./grocery-item.component.css'],
 })
 export class GroceryItemComponent implements OnInit {
-  @Input() listId!: number;
+  @Input() list!: GroceryList;
+  listId = this.list?.id;
+
+  selectedItem!: GroceryItem;
+
   groceryItems: GroceryItem[] = [];
-  newItemName: string = '';
-  newItemQuantity: number = 1;
+  groceryItemForm: FormGroup;
+  groceryItemFormModal: FormGroup;
 
   constructor(
     private route: ActivatedRoute,
-    private groceryService: SharedService
-  ) {}
+    private groceryService: SharedService,
+    private formBuilder: FormBuilder
+  ) {
+    this.groceryItemForm = this.formBuilder.group({
+      name: ['', Validators.required],
+      quantity: ['', Validators.required],
+    });
+
+    this.groceryItemFormModal = this.formBuilder.group({
+      name: ['', Validators.required],
+      quantity: ['', Validators.required],
+      purchased: [false],
+    });
+  }
 
   ngOnInit(): void {
+    this.loadGroceryItems();
+  }
+
+  loadGroceryItems(): void {
     this.route.paramMap
       .pipe(
         switchMap((params: ParamMap) => {
-          this.listId = +params.get('listId')!; // Get the listId parameter from the route
+          this.listId = +params.get('listId')!;
           return this.groceryService.getGroceryItemsForList(this.listId);
         })
       )
@@ -124,42 +172,78 @@ export class GroceryItemComponent implements OnInit {
       });
   }
 
-  createGroceryItem(name: string, quantity: number): void {
-    if (name.trim() === '') {
+  createGroceryItem(): void {
+    if (this.groceryItemForm.invalid) {
       return;
     }
 
     const newItem: Partial<GroceryItem> = {
-      name: name,
-      quantity: quantity,
-      // listId: this.listId,
+      name: this.groceryItemForm.value.name,
+      quantity: this.groceryItemForm.value.quantity,
+      grocery_list: this.listId,
     };
-
-    // Add the listId property separately
-    //newItem.listId = this.listId;
 
     this.groceryService
       .createGroceryItemForList(this.listId, newItem as GroceryItem)
       .subscribe((item) => {
         this.groceryItems.push(item);
-        this.newItemName = '';
-        this.newItemQuantity = 1;
+        this.groceryItemForm.reset(); // Reset the form
+        this.groceryItemForm.patchValue({ quantity: 1 }); // Set quantity to 1
       });
   }
 
-  updateGroceryItem(item: GroceryItem): void {
+  openModal(item: GroceryItem) {
+    this.selectedItem = item;
+    this.groceryItemFormModal.setValue({
+      name: item.name,
+      quantity: item.quantity,
+      purchased: item.purchased,
+    });
+  }
+
+  updateGroceryItem(): void {
+    if (!this.selectedItem) {
+      return;
+    }
+
+    // Update the selected grocery item with the edited values
+    const { name, quantity, purchased } = this.groceryItemFormModal.value;
+    this.selectedItem.name = name;
+    this.selectedItem.quantity = quantity;
+    this.selectedItem.purchased = purchased;
+
+    // Call the API to update the grocery item
     this.groceryService
-      .updateGroceryItemForList(this.listId, item.id, item)
+      .updateGroceryItemForList(
+        this.listId,
+        this.selectedItem.id,
+        this.selectedItem
+      )
       .subscribe(() => {
-        // You can handle any UI updates or notifications here
+        // After successful update, close the modal and reload the grocery items
+        this.loadGroceryItems();
       });
   }
 
   deleteGroceryItem(item: GroceryItem): void {
+    if (confirm('Are you sure?')) {
+      this.groceryService
+        .deleteGroceryItemForList(this.listId, item.id)
+        .subscribe(() => {
+          this.groceryItems = this.groceryItems.filter((i) => i.id !== item.id);
+        });
+    }
+  }
+
+  markAsPurchased(item: GroceryItem): void {
+    const purchasedValue = !item.purchased;
+    item.purchased = purchasedValue;
     this.groceryService
-      .deleteGroceryItemForList(this.listId, item.id)
+      .updateGroceryItemForList(this.listId, item.id, item)
       .subscribe(() => {
-        this.groceryItems = this.groceryItems.filter((i) => i.id !== item.id);
+        // If needed, you can reload the grocery items after the update
+        this.loadGroceryItems();
       });
+    //this.updateGroceryItem();
   }
 }
